+++ app/helpers/application_helper.rb
@@ -1,0 +1,4 @@
+module Weeler
+  module ApplicationHelper
+  end
+end