+++ lib/weeler/engine.rb
@@ -3,6 +3,8 @@ module Weeler
     config.railties_order = [:main_app, Weeler::Engine, :all]
     config.weeler = Weeler
 
+    config.i18n.available_locales = [:en] unless config.i18n.available_locales.present?
+
     config.assets.precompile += ["weeler/init.js", "weeler/init.css"]
   end