+++ spec/dummy/db/migrate/20140123083706_create_weeler_translations.rb
@@ -1,0 +1,16 @@
+class CreateWeelerTranslations < ActiveRecord::Migration
+  def change
+    create_table :weeler_translations do |t|
+      t.string :locale
+      t.string :key
+      t.text   :value
+      t.text   :interpolations
+      t.boolean :is_proc, :default => false
+
+      t.timestamps
+    end
+
+    add_index :weeler_translations, :locale
+    add_index :weeler_translations, :key
+  end
+end